engery_analysis_app/v1/backend/projects/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.1.0
Markdown==3.4.3
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.15.0
